Update __init__.py
AbstractDataStream requires that sources be a tuple. When sources are a list one get errors on concatenation with touples, e.g. when a FilterSources transformer is chained with a Mapping one (there is then a list to tuple concatenation at https://github.com/mila-udem/fuel/blob/master/fuel/transformers/__init__.py#L107-L108 )