update setup.py version and download_url
Reverted to using already defined version variable so
that it is not defined in two places. This earlier change
was an oversight.

Updated download_url to github v3 API url:

/:owner/:repo/:archive_format/:ref

Note for :ref, this scheme implies a future git tag
of vX.Y.Z for each released version X.Y.Z.