Reduces code duplication between ilsvrc2012/2010
Changes to ilsvrc2012:
As an approach to avoid any refactorings, I create empty patch images
archive and pass it to produce train/other set from ilsvrc2010.
The only functionality that is specific to ilsvrc2012 is split creation.

Changes to ilsvrc2010:
Before writing to hdf5 file check if class_index is provided. For
ilsvrc2012 for test set ground truth is not provided.